// src/app/api/models/current-rate.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */


/**
 * Everything the till needs about money in one call.
 */
export interface CurrentRate {
  base_currency: string;
  decimals: {
[key: string]: number;
};
  effective_date: string;
  notes: {
[key: string]: Array<string>;
};
  rate: string;
  rounding_step: {
[key: string]: string;
};
  symbol: {
[key: string]: string;
};
}