// src/app/api/models/me.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { RoleEnum } from '../models/role-enum';
export interface Me {
  full_name: string;
  full_name_kh?: string;
  id: number;
  must_change_password?: boolean;
  role?: RoleEnum;
  scopes: Array<string>;
  username: string;
}